Update Adafruit_FONA.cpp
I don't know if this might be an appropriate solution but I tried to replace the lines that gave me the errors reported in the following topic https://github.com/botletics/SIM7000-LTE-Shield/issues/110 and it seems to compile just fine on ESP32. I did not verify if these changes are actually working or not. Please verify them before updating.