Update Adafruit_FONA.cpp
Some generic eBay boards seem to use a different revision / firmware of the 808 chip that is detected slightly differently - its still an 808 chip, but has the bluetooth and other bits enabled. I will get round to working out some BT functions for it, and as your library is most improved from some of the other, I figured this is as good a start as any.