Remove some warnings so that it's easier to see real issues when doing a build.  Chunks of code are commented out when not doing specific things.
If you don't care about UTC time all of the serial print statements are
commented out but the declaration of the used variables are not commented
out generating an un-needed warning.

Likewise if you are not going to use the HTTP POST method to dweet.io the
code is commented out but the declaration of the variable "body" is not
commented out again generating an un-needed warning.